requirements.txt
pydantic>=2.0.0
openai>=1.12.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
praw>=7.7.1
beautifulsoup4>=4.12.0
requests>=2.31.0
serpapi>=0.1.0 